requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Flask==2.2.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
selenium==4.4.0
sniffio==1.2.0
sortedcontainers==2.4.0
stripe==4.0.2
trio==0.21.0
trio-websocket==0.9.2
twilio==7.12.0
urllib3==1.26.11
Werkzeug==2.2.2
wsproto==1.1.0
